+++ src/ast.rs
@@ -119,6 +119,7 @@ impl fmt::Debug for Func {
 pub struct Modified {
     pub modifier: Sp<Primitive>,
     pub operands: Vec<Sp<Word>>,
+    pub terminated: bool,
 }
 
 impl fmt::Debug for Modified {
@@ -126,6 +127,9 @@ impl fmt::Debug for Modified {
         write!(f, "{:?}", self.modifier.value)?;
         for word in &self.operands {
             write!(f, "({:?})", word.value)?;
+        }
+        if self.terminated {
+            write!(f, "|")?;
         }
         Ok(())
     }